Add checksum to transaction logs
Checksums are calculated over each transaction so we can validate the integrity of the stored files when doing recovery. CRC32C was chosen because it should theoretically give better performance when working in buffered reader/writer as we do.

Also, remove masterId and authorId from transactions since that is no longer the source for checksums.

+++ private/enterprise/enterprise-it/kernel-it/src/test/java/com/neo4j/internal/batchimport/store/BatchingNeoStoresIT.java
@@ -106,7 +106,7 @@ class BatchingNeoStoresIT
         }
 
         @Override
-        public long lastCommittedTransactionChecksum()
+        public int lastCommittedTransactionChecksum()
         {
             return 11;
         }
